ci: add install gsasl
