+++ .github/workflows/ci.yaml
@@ -9,6 +9,8 @@ jobs:
   check:
     runs-on: ubuntu-latest
     steps:
+      - name: Install gsasl
+        run: sudo apt-get install -y libgsasl-dev
       - uses: actions/checkout@v2
       - uses: actions/cache@v1
         with:
