+++ vsmtp-config/Cargo.toml
@@ -18,8 +18,6 @@ vsmtp-common = { version = "^0.9", path = "../vsmtp-common" }
 log = "0.4.14"
 log4rs = { version = "1.0.0" }
 
-anyhow = "1.0.56"
-
 users = { version = "0.11.0", features = [] }
 
 serde = { version = "1.0.136", features = ["derive"] }
