+++ vsmtp-config/src/parser/tls_certificate.rs
@@ -1,3 +1,5 @@
+use vsmtp_common::re::anyhow;
+
 pub(crate) fn from_string(input: &str) -> anyhow::Result<rustls::Certificate> {
     let path = std::path::Path::new(&input);
     anyhow::ensure!(
