+++ vsmtp-config/src/rustls_helper.rs
@@ -1,3 +1,5 @@
+use vsmtp_common::re::anyhow;
+
 use crate::config::ConfigServerTls;
 
 struct TlsLogger;
