+++ vsmtp/Cargo.toml
@@ -17,12 +17,10 @@ vsmtp-common = { version = "^0.9", path = "../vsmtp-common" }
 vsmtp-server = { version = "^0.9", path = "../vsmtp-server" }
 vsmtp-config = { version = "^0.9", path = "../vsmtp-config" }
 
-log = "0.4.14"
+log = "0.4.16"
 log4rs = { version = "1.0.0" }
 
 clap = { version = "3.1.6", features = ["derive"] }
-
-anyhow = "1.0.56"
 
 serde_json = "1.0.79"
