+++ vsmtp/src/lib.rs
@@ -9,6 +9,8 @@
 #![warn(clippy::cargo)]
 //
 #![allow(clippy::doc_markdown)]
+#![allow(clippy::multiple_crate_versions)]
 
-/// Argument for the execution of vSMTP
-pub mod args;
+mod args;
+
+pub use args::{Args, Commands};
